fix: use better comparison for benchmark, # of runs
- write result to json for egress to future internal benchmark service
- sample threads to get better data

+++ Samples/TrendingMovies/TrendingMovies/SentryBenchmarking.h
@@ -5,7 +5,7 @@ NS_ASSUME_NONNULL_BEGIN
 @interface SentryBenchmarking : NSObject
 
 + (void)startBenchmarkProfile;
-+ (uint64_t)retrieveBenchmarks;
++ (double)retrieveBenchmarks;
 
 @end
